LayoutTests/websocket/tests/bad-sub-protocol-expected.txt

CONSOLE MESSAGE: line 0: Wrong protocol for WebSocket 
CONSOLE MESSAGE: line 0: Wrong protocol for WebSocket 
CONSOLE MESSAGE: line 0: Wrong protocol for WebSocket 	
CONSOLE MESSAGE: line 0: Wrong protocol for WebSocket 
CONSOLE MESSAGE: line 0: Wrong protocol for WebSocket 
CONSOLE MESSAGE: line 0: Wrong protocol for WebSocket 
CONSOLE MESSAGE: line 0: Wrong protocol for WebSocket あ
CONSOLE MESSAGE: line 0: Wrong protocol for WebSocket ￿
CONSOLE MESSAGE: line 0: Wrong protocol for WebSocket 
CONSOLE MESSAGE: line 0: Wrong protocol for WebSocket ￾
CONSOLE MESSAGE: line 0: Wrong protocol for WebSocket 𠀋
Test WebSocket bad sub-protocol names.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS new WebSocket("ws://127.0.0.1:8880/simple", "") threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "") threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "	") threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "
") threw exception SyntaxError: Parse error.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "
") threw exception SyntaxError: Parse error.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "") threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "") threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "") threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "あ") threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "￿") threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "﻿") threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "￾") threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "𠀋") threw exception Error: SYNTAX_ERR: DOM Exception 12.
